clean: refactor & use dotenv
