--- env.yaml
@@ -1,66 +1,0 @@
-name: calendar
-channels:
-  - conda-forge
-  - defaults
-dependencies:
-  - _libgcc_mutex=0.1=main
-  - _openmp_mutex=5.1=1_gnu
-  - blas=1.0=mkl
-  - bottleneck=1.3.5=py39h7deecbd_0
-  - ca-certificates=2022.6.15=ha878542_0
-  - certifi=2022.6.15=py39hf3d152e_0
-  - intel-openmp=2021.4.0=h06a4308_3561
-  - ld_impl_linux-64=2.38=h1181459_1
-  - libffi=3.3=he6710b0_2
-  - libgcc-ng=11.2.0=h1234567_1
-  - libgomp=11.2.0=h1234567_1
-  - libstdcxx-ng=11.2.0=h1234567_1
-  - mkl=2021.4.0=h06a4308_640
-  - mkl-service=2.4.0=py39h7e14d7c_0
-  - mkl_fft=1.3.1=py39h0c7bc48_1
-  - mkl_random=1.2.2=py39hde0f152_0
-  - ncurses=6.3=h5eee18b_3
-  - numexpr=2.8.3=py39h807cd23_0
-  - numpy=1.23.1=py39h6c91a56_0
-  - numpy-base=1.23.1=py39ha15fc14_0
-  - openssl=1.1.1q=h7f8727e_0
-  - packaging=21.3=pyhd8ed1ab_0
-  - pandas=1.4.3=py39h6a678d5_0
-  - pip=22.1.2=py39h06a4308_0
-  - pyparsing=3.0.9=pyhd8ed1ab_0
-  - python=3.9.12=h12debd9_1
-  - python-dateutil=2.8.2=pyhd8ed1ab_0
-  - python_abi=3.9=2_cp39
-  - pytz=2022.1=pyhd8ed1ab_0
-  - readline=8.1.2=h7f8727e_1
-  - setuptools=61.2.0=py39h06a4308_0
-  - six=1.16.0=pyh6c4a22f_0
-  - sqlite=3.39.2=h5082296_0
-  - tk=8.6.12=h1ccaba5_0
-  - tzdata=2022a=hda174b7_0
-  - wheel=0.37.1=pyhd3eb1b0_0
-  - xz=5.2.5=h7f8727e_1
-  - zlib=1.2.12=h7f8727e_2
-  - pip:
-      - cachetools==5.2.0
-      - charset-normalizer==2.1.0
-      - google-api-core==2.8.2
-      - google-api-python-client==2.55.0
-      - google-auth==2.10.0
-      - google-auth-httplib2==0.1.0
-      - google-auth-oauthlib==0.5.2
-      - googleapis-common-protos==1.56.4
-      - httplib2==0.20.4
-      - idna==3.3
-      - oauthlib==3.2.0
-      - protobuf==4.21.4
-      - pyasn1==0.4.8
-      - pyasn1-modules==0.2.8
-      - requests==2.28.1
-      - requests-oauthlib==1.3.1
-      - rsa==4.9
-      - slack-bolt==1.14.3
-      - slack-sdk==3.18.1
-      - uritemplate==4.1.1
-      - urllib3==1.26.11
-prefix: /home/usrs/mitsui/miniconda3/envs/calendar
